requirements.txt
openai
chromadb
pypdf
python-dotenv
pytest
hypothesis
boto3
